Refactor index tests:
- Factor out `forAllFixtures` function, to run a test over all the storage types.
- Move some utils functions to pkg/chunk/storage/utils_test.go.
